+++ limen/sfd/foundational_sfd/random_binary.py
@@ -30,11 +30,11 @@ def manifest():
             'datetime', 'high', 'low', 'close', 'volume', 'maker_ratio',
             'no_of_trades'
         ])
-        .with_target('outcome')
+        .with_target_label('outcome')
             .add_transform(lambda data: data.with_columns(
                 pl.Series('outcome', np.random.randint(0, 2, size=data.height))
             ))
             .add_transform(lambda data: data[:-1])
             .done()
-        .with_model(random_binary)
+        .with_reference_architecture(random_binary)
     )