+++ tests/test_manifest_pre_split_random_selector.py
@@ -23,7 +23,7 @@ def test_pre_split_random_selector():
             seed='random_seed'
         )
         .set_split_config(6, 2, 2)
-        .with_target('target')
+        .with_target_label('target')
             .done()
     )